add: EstimationStudy()
